Update revamped.cpp
I've tried to revamp this project to the best of my ability with the little cpp knowledge I have. I have also tried to tidy it up with my C knowledge but I'm afraid this will not suffice, I will return to this project when I have more cpp knowledge. I want this project to be robust and eventually have a Graphical user interface, this is it for now.